Refactor to make functions more modular and resolve cyclic dependency
