+++ src/index.js
@@ -1,18 +1,18 @@
 import './style.css';
 import Refresh from './rotate.svg';
 import Enter from './enter.png';
-import Todo from './modules/todo.js';
-import { createElement, appendToDOM } from './modules/handleDOM.js';
+import { appendToDOM } from './modules/handleDOM.js';
+import createShowElement from './modules/showTodo.js';
 import { retrieveData } from './modules/handleData.js';
-
-let todoArray = [];
+import { newTodoForm, newTodoFormHandler } from './modules/newTodoForm.js';
 
 const loadElements = () => {
+  let todoArray = [];
   todoArray = retrieveData();
   todoArray
     .sort((a, b) => a.index - b.index)
     .forEach((todo) => {
-      const todoElement = createElement(todo);
+      const todoElement = createShowElement(todo);
       appendToDOM(todoElement);
     });
 };
@@ -35,21 +35,5 @@ window.onload = () => {
   loadElements();
   loadRefreshBtn();
   loadEnterBtn();
+  newTodoForm.addEventListener('submit', newTodoFormHandler);
 };
-
-const todoForm = document.getElementById('todo-form');
-
-const todoFormHandler = (e) => {
-  e.preventDefault();
-  const newTodo = new Todo(
-    todoForm.elements['add-todo'].value,
-    false,
-    new Date().valueOf(),
-  );
-  Todo.addTodo(newTodo);
-  const todoElement = createElement(newTodo);
-  appendToDOM(todoElement);
-  todoForm.reset();
-};
-
-todoForm.addEventListener('submit', todoFormHandler);
